Revert globToRegex as fs.globSync won't work for proxy in the future
