+++ lib/file_util.js
@@ -4,7 +4,7 @@ const fsPromises = require('node:fs/promises');
 const path = require('path');
 const zlib = require('zlib');
 const os = require('os');
-const { isWindows, globToRegex } = require('./util');
+const { isWindows, escapeRegex } = require('./util');
 const Logger = require('./logger').default;
 
 const resultStatus = {
@@ -135,8 +135,27 @@ function FileUtil() {
 }
 exports.FileUtil = FileUtil;
 
+// NOTE:
+// This won't be needed when we'll support Node 22+ which has fs.globSync method
+exports.globToRegex = function (pattern) {
+  let regexPattern = '';
+
+  for (let i = 0; i < pattern.length; i++) {
+    const char = pattern[i];
+    if (char === '*') {
+      regexPattern += '.*';
+    } else if (char === '?') {
+      regexPattern += '.';
+    } else {
+      regexPattern += escapeRegex(char);
+    }
+  }
+
+  return new RegExp(`^${regexPattern}$`);
+};
+
 exports.getMatchingFilePaths = function (dir, fileName) {
-  const fileNameRegex = globToRegex(fileName);
+  const fileNameRegex = exports.globToRegex(fileName);
   try {
     const files = fs.readdirSync(dir);
     const matchedFiles = files.filter((file) => fileNameRegex.test(file));
