+++ lib/proxy_util.js
@@ -347,14 +347,20 @@ exports.isByPassProxy = function (proxy, destination) {
   if (proxy && proxy.noProxy) {
     const bypassList = proxy.noProxy.split('|');
     return bypassList.find((noProxy) => {
-      let host = noProxy.trim();
+      const host = noProxy.trim();
       if (destination instanceof RegExp) {
         return destination.test(host);
       }
-      if (host.startsWith('.')) {
-        host = '*' + host;
+      let regexPattern = '';
+      for (let i = 0; i < host.length; i++) {
+        const char = host[i];
+        if (char === '*' || (char === '.' && i === 0)) {
+          regexPattern += '.*';
+        } else {
+          regexPattern += Util.escapeRegex(char);
+        }
       }
-      const noProxyRegex = Util.globToRegex(host);
+      const noProxyRegex = new RegExp(`^${regexPattern}$`);
       return noProxyRegex.test(destination);
     });
   }
